Make peer_actor.rs smaller and therefore easier to read
