+++ chain/network/src/peer/mod.rs
@@ -1,2 +1,3 @@
 pub mod peer_actor;
 mod rate_counter;
+mod tracker;
